requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow
streamlit
pillow
argparse
